# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test IBID API",
        "contact": {
            "name": "Budi Haryono",
            "email": "budiharyono4@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication (Soal no. 2)"
                ],
                "summary": "Sign in",
                "description": "Login by email, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Email and Password Credential",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "asdf@zxcv.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123asd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return JWT Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "auth_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0OjgwMDAvYXBpL2xvZ2luIiwiaWF0IjoxNjI2NTk2MjQ5LCJleHAiOjE2MjkyMjQyNDksIm5iZiI6MTYyNjU5NjI0OSwianRpIjoiWGlrczhCQjJrbGdxSkFyTiIsInN1YiI6IjYwZjJhODM3OTZlYTczMWJkZjA4ZmNmOSIsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.pFNfya7fcSXkyuOZ-obxSyJLSdt8uiYAvKkYiHEEIrg"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "string",
                                            "example": "43800"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credentials not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Email or Password not found"
                                        },
                                        "result": {
                                            "type": "null",
                                            "example": "null"
                                        },
                                        "error_code": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Input validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "result": {
                                            "type": "null",
                                            "example": "null"
                                        },
                                        "error_code": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "field": {
                                                        "type": "string",
                                                        "example": "The email field is required."
                                                    },
                                                    "message": {
                                                        "type": "string",
                                                        "example": "The password field is required."
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "field": "email",
                                                    "message": "The email input is required."
                                                },
                                                {
                                                    "field": "password",
                                                    "message": "The email input is required."
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication (Soal no. 2)"
                ],
                "summary": "Register",
                "description": "Register using email, name, password",
                "operationId": "authRegister",
                "requestBody": {
                    "description": "Email and Password Credential",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "name",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "budi@domain.com"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Budi Haryono"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123asd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return JWT Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "auth_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0OjgwMDAvYXBpL2xvZ2luIiwiaWF0IjoxNjI2NTk2MjQ5LCJleHAiOjE2MjkyMjQyNDksIm5iZiI6MTYyNjU5NjI0OSwianRpIjoiWGlrczhCQjJrbGdxSkFyTiIsInN1YiI6IjYwZjJhODM3OTZlYTczMWJkZjA4ZmNmOSIsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.pFNfya7fcSXkyuOZ-obxSyJLSdt8uiYAvKkYiHEEIrg"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "string",
                                            "example": "43800"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Input validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "result": {
                                            "type": "null",
                                            "example": "null"
                                        },
                                        "error_code": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "field": {
                                                        "type": "string",
                                                        "example": "The email field is required."
                                                    },
                                                    "message": {
                                                        "type": "string",
                                                        "example": "The password field is required."
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "field": "email",
                                                    "message": "The email input is required."
                                                },
                                                {
                                                    "field": "password",
                                                    "message": "The email input is required."
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication (Soal no. 2)"
                ],
                "summary": "Logout",
                "description": "Logout user and invalidate token",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}